moved stuff into folders, added nlprg yaml
